refactor(admin-ui): extract Thumbnail into its own file
